WIP __CSPHandleMIDIMsg

+++ include/audio_internal.h
@@ -296,7 +296,7 @@ typedef struct PVoice_s {
     ALResampler         resampler;
     ALEnvMixer		envmixer;
     s32                 offset;
-    s32                 unkDC;
+    u8                 unkDC;
 } PVoice;
 
 #endif
